pyproject.toml
[tool.poetry]
name = "repl_python3_Tameshk"
version = "0.1.0"
description = ""
authors = ["Hadi Azarnasab <hazard7546@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.17.1"
discord = "^1.0.1"
pyautogui = "^0.9.52"
Khayyam = "^3.0.17"
randomdict = "^0.2.0"
googletrans = "^4.0.0-rc1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
